% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_codes}
\alias{get_codes}
\title{Get all codes and metadata for WHO series}
\usage{
get_codes(extra = FALSE)
}
\arguments{
\item{extra}{logical If \code{TRUE}, downloads additional meta
information (e.g. series categories, French and Spanish descriptions
(where available), etc).}
}
\value{
A data frame
}
\description{
Get all codes and metadata for WHO series
}
\examples{
codes <- get_codes()
str(codes)
}

