% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_data}
\alias{get_data}
\title{Retrieve Data from the World Health Organization}
\usage{
get_data(code)
}
\arguments{
\item{code}{character The code for the time series to be retrieved

Time-series codes can be retrieved through the \code{get_codes} function.}
}
\value{
A data frame
}
\description{
Retrieve Data from the World Health Organization
}
\examples{
df <- get_data("WHOSIS_000001")
head(df)
}

