% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-data.R
\name{get_who_data}
\alias{get_who_data}
\title{get_who_data}
\usage{
get_who_data(city = "kathmandu", key = NULL, value = NULL)
}
\arguments{
\item{city}{Name of city for which data are to be obtained}

\item{key}{List of OSM keys}

\item{value}{List of corresponding OSM values}
}
\value{
The \pkg{sf}-formatted data object (invisibly)
}
\description{
Extract OSM data for given location (\code{city}), and save it in the data
directory
}
\note{
\code{value} must either be \code{NULL}, or have the same length as
\code{key}. To return all values for one particular \code{key} from a vector,
set corresponding \code{value = ""}.
}
