% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-data.R
\name{get_who_streets}
\alias{get_who_streets}
\title{get_who_streets}
\usage{
get_who_streets(city = "kathmandu")
}
\arguments{
\item{city}{Name of city for which streets are to be obtained}
}
\value{
The \pkg{sf}-formatted data object (invisibly)
}
\description{
Extract OSM streets for given location (\code{city}), and save them in the
data directory
}
