% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards.R
\docType{data}
\name{standards}
\alias{standards}
\title{Standards data from who}
\format{a tibble with columns
- Age: Age in days
- sex: F or M
- percentile: percentile value
- weight: weight in kg
- height: height in cm
- head_circumference: head circumference in cm
- arm_circumference: arm circumference in cm
- subscapular_skinfold
- triceps_skinfold}
\usage{
standards
}
\description{
Standards data from who
}
\keyword{datasets}
